factor out separable problem into its own module
