+++ benches/double_exponential_without_noise.rs
@@ -13,8 +13,8 @@ use shared_test_code::models::DoubleExpModelWithConstantOffsetSepModel;
 use shared_test_code::models::DoubleExponentialDecayFittingWithOffsetLevmar;
 use shared_test_code::*;
 use varpro::prelude::SeparableNonlinearModel;
+use varpro::problem::SeparableProblem;
 use varpro::solvers::levmar::LevMarSolver;
-use varpro::solvers::levmar::SeparableProblem;
 use varpro::solvers::levmar::SeparableProblemBuilder;
 
 /// helper struct for the parameters of the double exponential
